requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11
django-appconf==1.0.2
django-crispy-forms==1.7.0
django-downloadview==1.9
django-reversion==2.0.12
django-widget-tweaks==1.4.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
lxml==4.2.1
Pillow==5.1.0
PyPDF2==1.26.0
python-docx==0.8.6
pytz==2018.4
reportlab==3.4.0
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
webencodings==0.5.1
